# Use FastHTTP client to improve HTTP performances 