--- a/cmd/configuration.go
+++ b/cmd/configuration.go
@@ -25,9 +25,6 @@ func NewTraefikConfiguration() *TraefikCmdConfiguration {
 			Providers: &static.Providers{
 				ProvidersThrottleDuration: ptypes.Duration(2 * time.Second),
 			},
-			ServersTransport: &static.ServersTransport{
-				MaxIdleConnsPerHost: 200,
-			},
 			TCPServersTransport: &static.TCPServersTransport{
 				DialTimeout:   ptypes.Duration(30 * time.Second),
 				DialKeepAlive: ptypes.Duration(15 * time.Second),