--- a/cmd/traefik/traefik.go
+++ b/cmd/traefik/traefik.go
@@ -36,14 +36,14 @@ import (
 	"github.com/traefik/traefik/v3/pkg/provider/aggregator"
 	"github.com/traefik/traefik/v3/pkg/provider/tailscale"
 	"github.com/traefik/traefik/v3/pkg/provider/traefik"
-	"github.com/traefik/traefik/v2/pkg/proxy"
+	"github.com/traefik/traefik/v3/pkg/proxy"
 	"github.com/traefik/traefik/v3/pkg/safe"
 	"github.com/traefik/traefik/v3/pkg/server"
 	"github.com/traefik/traefik/v3/pkg/server/middleware"
 	"github.com/traefik/traefik/v3/pkg/server/service"
 	"github.com/traefik/traefik/v3/pkg/tcp"
 	traefiktls "github.com/traefik/traefik/v3/pkg/tls"
-	"github.com/traefik/traefik/v2/pkg/tls/client"
+	"github.com/traefik/traefik/v3/pkg/tls/client"
 	"github.com/traefik/traefik/v3/pkg/tracing"
 	"github.com/traefik/traefik/v3/pkg/tracing/jaeger"
 	"github.com/traefik/traefik/v3/pkg/types"
@@ -261,9 +261,14 @@ func setupServer(staticConfiguration *static.Configuration) (*server.Server, err
 		log.Info().Msg("Successfully obtained SPIFFE SVID.")
 	}
 
-	dialerManager := tcp.NewDialerManager(spiffeX509Source)
-	tlsClientConfigManager := client.NewTLSConfigManager(spiffeX509Source)
+	// TCP Manager
+	tcpTLSClientConfigManager := client.NewTLSConfigManager[*dynamic.TCPServersTransport](spiffeX509Source)
+	dialerManager := tcp.NewDialerManager(tcpTLSClientConfigManager)
+
+	// HTTP Manager
+	tlsClientConfigManager := client.NewTLSConfigManager[*dynamic.ServersTransport](spiffeX509Source)
 	proxyBuilder := proxy.NewBuilder(tlsClientConfigManager)
+
 	acmeHTTPHandler := getHTTPChallengeHandler(acmeProviders, httpChallengeProvider)
 	managerFactory := service.NewManagerFactory(*staticConfiguration, routinesPool, metricsRegistry, proxyBuilder, tlsClientConfigManager, acmeHTTPHandler)
 
@@ -304,6 +309,8 @@ func setupServer(staticConfiguration *static.Configuration) (*server.Server, err
 	// Server Transports
 	watcher.AddListener(func(conf dynamic.Configuration) {
 		tlsClientConfigManager.Update(conf.HTTP.ServersTransports)
+		tcpTLSClientConfigManager.Update(conf.TCP.ServersTransports)
+
 		proxyBuilder.Update(conf.HTTP.ServersTransports)
 		dialerManager.Update(conf.TCP.ServersTransports)
 	})