--- a/pkg/config/dynamic/http_config.go
+++ b/pkg/config/dynamic/http_config.go
@@ -15,11 +15,11 @@ const (
 	// DefaultHealthCheckTimeout is the default value for the ServerHealthCheck timeout.
 	DefaultHealthCheckTimeout = ptypes.Duration(5 * time.Second)
 
-	// DefaultPassHostHeader is the default value for the ServersLoadBalancer passHostHeader.
+	// DefaultPassHostHeader is the default value for the HTTPClientConfig passHostHeader.
 	DefaultPassHostHeader = true
 
-	// DefaultFlushInterval is the default value for the ResponseForwarding flush interval.
-	DefaultFlushInterval = ptypes.Duration(100 * time.Millisecond)
+	// DefaultIdleConnTimeout is the default value for the ForwardingTimeouts idleConnTimeout.
+	DefaultIdleConnTimeout = ptypes.Duration(90 * time.Second)
 )
 
 // +k8s:deepcopy-gen=true
@@ -167,10 +167,8 @@ type ServersLoadBalancer struct {
 	// children servers of this load-balancer. To propagate status changes (e.g. all
 	// servers of this service are down) upwards, HealthCheck must also be enabled on
 	// the parent(s) of this service.
-	HealthCheck        *ServerHealthCheck  `json:"healthCheck,omitempty" toml:"healthCheck,omitempty" yaml:"healthCheck,omitempty" export:"true"`
-	PassHostHeader     *bool               `json:"passHostHeader" toml:"passHostHeader" yaml:"passHostHeader" export:"true"`
-	ResponseForwarding *ResponseForwarding `json:"responseForwarding,omitempty" toml:"responseForwarding,omitempty" yaml:"responseForwarding,omitempty" export:"true"`
-	ServersTransport   string              `json:"serversTransport,omitempty" toml:"serversTransport,omitempty" yaml:"serversTransport,omitempty" export:"true"`
+	HealthCheck      *ServerHealthCheck `json:"healthCheck,omitempty" toml:"healthCheck,omitempty" yaml:"healthCheck,omitempty" export:"true"`
+	ServersTransport string             `json:"serversTransport,omitempty" toml:"serversTransport,omitempty" yaml:"serversTransport,omitempty" export:"true"`
 }
 
 // Mergeable tells if the given service is mergeable.
@@ -188,32 +186,6 @@ func (l *ServersLoadBalancer) Mergeable(loadBalancer *ServersLoadBalancer) bool 
 	loadBalancer.Servers = nil
 
 	return reflect.DeepEqual(l, loadBalancer)
-}
-
-// SetDefaults Default values for a ServersLoadBalancer.
-func (l *ServersLoadBalancer) SetDefaults() {
-	defaultPassHostHeader := DefaultPassHostHeader
-	l.PassHostHeader = &defaultPassHostHeader
-
-	l.ResponseForwarding = &ResponseForwarding{}
-	l.ResponseForwarding.SetDefaults()
-}
-
-// +k8s:deepcopy-gen=true
-
-// ResponseForwarding holds the response forwarding configuration.
-type ResponseForwarding struct {
-	// FlushInterval defines the interval, in milliseconds, in between flushes to the client while copying the response body.
-	// A negative value means to flush immediately after each write to the client.
-	// This configuration is ignored when ReverseProxy recognizes a response as a streaming response;
-	// for such responses, writes are flushed to the client immediately.
-	// Default: 100ms
-	FlushInterval ptypes.Duration `json:"flushInterval,omitempty" toml:"flushInterval,omitempty" yaml:"flushInterval,omitempty" export:"true"`
-}
-
-// SetDefaults Default values for a ResponseForwarding.
-func (r *ResponseForwarding) SetDefaults() {
-	r.FlushInterval = DefaultFlushInterval
 }
 
 // +k8s:deepcopy-gen=true
@@ -263,17 +235,51 @@ type HealthCheck struct{}
 
 // +k8s:deepcopy-gen=true
 
-// ServersTransport options to configure communication between Traefik and the servers.
+// ServersTransport holds the configuration of the communication between Traefik and the servers.
 type ServersTransport struct {
-	ServerName          string                     `description:"Defines the serverName used to contact the server." json:"serverName,omitempty" toml:"serverName,omitempty" yaml:"serverName,omitempty"`
-	InsecureSkipVerify  bool                       `description:"Disables SSL certificate verification." json:"insecureSkipVerify,omitempty" toml:"insecureSkipVerify,omitempty" yaml:"insecureSkipVerify,omitempty" export:"true"`
-	RootCAs             []traefiktls.FileOrContent `description:"Defines a list of CA secret used to validate self-signed certificate" json:"rootCAs,omitempty" toml:"rootCAs,omitempty" yaml:"rootCAs,omitempty"`
-	Certificates        traefiktls.Certificates    `description:"Defines a list of secret storing client certificates for mTLS." json:"certificates,omitempty" toml:"certificates,omitempty" yaml:"certificates,omitempty" export:"true"`
-	MaxIdleConnsPerHost int                        `description:"If non-zero, controls the maximum idle (keep-alive) to keep per-host. If zero, DefaultMaxIdleConnsPerHost is used" json:"maxIdleConnsPerHost,omitempty" toml:"maxIdleConnsPerHost,omitempty" yaml:"maxIdleConnsPerHost,omitempty" export:"true"`
-	ForwardingTimeouts  *ForwardingTimeouts        `description:"Defines the timeouts for requests forwarded to the backend servers." json:"forwardingTimeouts,omitempty" toml:"forwardingTimeouts,omitempty" yaml:"forwardingTimeouts,omitempty" export:"true"`
-	DisableHTTP2        bool                       `description:"Disables HTTP/2 for connections with backend servers." json:"disableHTTP2,omitempty" toml:"disableHTTP2,omitempty" yaml:"disableHTTP2,omitempty" export:"true"`
-	PeerCertURI         string                     `description:"Defines the URI used to match against SAN URI during the peer certificate verification." json:"peerCertURI,omitempty" toml:"peerCertURI,omitempty" yaml:"peerCertURI,omitempty" export:"true"`
-	Spiffe              *Spiffe                    `description:"Defines the SPIFFE configuration." json:"spiffe,omitempty" toml:"spiffe,omitempty" yaml:"spiffe,omitempty" label:"allowEmpty" file:"allowEmpty" export:"true"`
+	HTTP *HTTPClientConfig `json:"http,omitempty" toml:"http,omitempty" yaml:"http,omitempty" export:"true"`
+	TLS  *TLSClientConfig  `json:"tls,omitempty" toml:"tls,omitempty" yaml:"tls,omitempty" export:"true"`
+}
+
+func (s *ServersTransport) SetDefaults() {
+	s.HTTP = &HTTPClientConfig{}
+	s.HTTP.SetDefaults()
+}
+
+// +k8s:deepcopy-gen=true
+
+// HTTPClientConfig holds the HTTP configuration to be used between Traefik and the servers.
+type HTTPClientConfig struct {
+	// PassHostHeader defines whether to forward the client Host header to the server.
+	PassHostHeader bool `json:"passHostHeader,omitempty" toml:"passHostHeader,omitempty" yaml:"passHostHeader,omitempty" export:"true"`
+	// MaxIdleConnsPerHost controls the maximum number of idle (keep-alive) connections to keep per-host.
+	// If zero, DefaultMaxIdleConnsPerHost is used.
+	MaxIdleConnsPerHost int `json:"maxIdleConnsPerHost,omitempty" toml:"maxIdleConnsPerHost,omitempty" yaml:"maxIdleConnsPerHost,omitempty" export:"true"`
+	// ForwardingTimeouts defines the timeouts for the requests forwarded to the backend servers.
+	ForwardingTimeouts *ForwardingTimeouts `json:"forwardingTimeouts,omitempty" toml:"forwardingTimeouts,omitempty" yaml:"forwardingTimeouts,omitempty" export:"true"`
+	// EnableHTTP2 enables HTTP/2 between Traefik and the backends.
+	EnableHTTP2 bool `json:"enableHTTP2,omitempty" toml:"enableHTTP2,omitempty" yaml:"enableHTTP2,omitempty" export:"true"`
+}
+
+// SetDefaults sets the default HTTPClientConfig values.
+func (h *HTTPClientConfig) SetDefaults() {
+	h.PassHostHeader = DefaultPassHostHeader
+	h.MaxIdleConnsPerHost = 200
+	h.ForwardingTimeouts = &ForwardingTimeouts{}
+	h.ForwardingTimeouts.SetDefaults()
+}
+
+// +k8s:deepcopy-gen=true
+
+// TLSClientConfig holds the TLS configuration to be used between Traefik and the servers.
+type TLSClientConfig struct {
+	ServerName         string                     `description:"Defines the serverName used to contact the server." json:"serverName,omitempty" toml:"serverName,omitempty" yaml:"serverName,omitempty"`
+	InsecureSkipVerify bool                       `description:"Disables SSL certificate verification." json:"insecureSkipVerify,omitempty" toml:"insecureSkipVerify,omitempty" yaml:"insecureSkipVerify,omitempty" export:"true"`
+	RootCAs            []traefiktls.FileOrContent `description:"Defines a list of CA secret used to validate self-signed certificate" json:"rootCAs,omitempty" toml:"rootCAs,omitempty" yaml:"rootCAs,omitempty"`
+	Certificates       traefiktls.Certificates    `description:"Defines a list of secret storing client certificates for mTLS." json:"certificates,omitempty" toml:"certificates,omitempty" yaml:"certificates,omitempty" export:"true"`
+	DisableHTTP2       bool                       `description:"Disables HTTP/2 for connections with backend servers." json:"disableHTTP2,omitempty" toml:"disableHTTP2,omitempty" yaml:"disableHTTP2,omitempty" export:"true"`
+	PeerCertURI        string                     `description:"Defines the URI used to match against SAN URI during the peer certificate verification." json:"peerCertURI,omitempty" toml:"peerCertURI,omitempty" yaml:"peerCertURI,omitempty" export:"true"`
+	Spiffe             *Spiffe                    `description:"Defines the SPIFFE configuration." json:"spiffe,omitempty" toml:"spiffe,omitempty" yaml:"spiffe,omitempty" label:"allowEmpty" file:"allowEmpty" export:"true"`
 }
 
 // +k8s:deepcopy-gen=true
@@ -281,25 +287,33 @@ type ServersTransport struct {
 // Spiffe holds the SPIFFE configuration.
 type Spiffe struct {
 	// IDs defines the allowed SPIFFE IDs (takes precedence over the SPIFFE TrustDomain).
-	IDs []string `description:"Defines the allowed SPIFFE IDs (takes precedence over the SPIFFE TrustDomain)." json:"ids,omitempty" toml:"ids,omitempty" yaml:"ids,omitempty"`
+	IDs []string `json:"ids,omitempty" toml:"ids,omitempty" yaml:"ids,omitempty"`
 	// TrustDomain defines the allowed SPIFFE trust domain.
-	TrustDomain string `description:"Defines the allowed SPIFFE trust domain." json:"trustDomain,omitempty" toml:"trustDomain,omitempty" yaml:"trustDomain,omitempty"`
+	TrustDomain string `json:"trustDomain,omitempty" toml:"trustDomain,omitempty" yaml:"trustDomain,omitempty"`
 }
 
 // +k8s:deepcopy-gen=true
 
-// ForwardingTimeouts contains timeout configurations for forwarding requests to the backend servers.
+// ForwardingTimeouts contains the timeout configurations for forwarding requests to the backend servers.
 type ForwardingTimeouts struct {
-	DialTimeout           ptypes.Duration `description:"The amount of time to wait until a connection to a backend server can be established. If zero, no timeout exists." json:"dialTimeout,omitempty" toml:"dialTimeout,omitempty" yaml:"dialTimeout,omitempty" export:"true"`
-	ResponseHeaderTimeout ptypes.Duration `description:"The amount of time to wait for a server's response headers after fully writing the request (including its body, if any). If zero, no timeout exists." json:"responseHeaderTimeout,omitempty" toml:"responseHeaderTimeout,omitempty" yaml:"responseHeaderTimeout,omitempty" export:"true"`
-	IdleConnTimeout       ptypes.Duration `description:"The maximum period for which an idle HTTP keep-alive connection will remain open before closing itself." json:"idleConnTimeout,omitempty" toml:"idleConnTimeout,omitempty" yaml:"idleConnTimeout,omitempty" export:"true"`
-	ReadIdleTimeout       ptypes.Duration `description:"The timeout after which a health check using ping frame will be carried out if no frame is received on the HTTP/2 connection. If zero, no health check is performed." json:"readIdleTimeout,omitempty" toml:"readIdleTimeout,omitempty" yaml:"readIdleTimeout,omitempty" export:"true"`
-	PingTimeout           ptypes.Duration `description:"The timeout after which the HTTP/2 connection will be closed if a response to ping is not received." json:"pingTimeout,omitempty" toml:"pingTimeout,omitempty" yaml:"pingTimeout,omitempty" export:"true"`
+	// DialTimeout defines the amount of time to wait until a connection to a backend server can be established.
+	// If zero, no timeout exists.
+	DialTimeout ptypes.Duration `json:"dialTimeout,omitempty" toml:"dialTimeout,omitempty" yaml:"dialTimeout,omitempty" export:"true"`
+	// ResponseHeaderTimeout defines the amount of time to wait for a server's response headers after fully writing the request (including its body, if any).
+	// If zero, no timeout exists.
+	ResponseHeaderTimeout ptypes.Duration `json:"responseHeaderTimeout,omitempty" toml:"responseHeaderTimeout,omitempty" yaml:"responseHeaderTimeout,omitempty" export:"true"`
+	// IdleConnTimeout defines the maximum period for which an idle HTTP keep-alive connection will remain open before closing itself.
+	IdleConnTimeout ptypes.Duration `json:"idleConnTimeout,omitempty" toml:"idleConnTimeout,omitempty" yaml:"idleConnTimeout,omitempty" export:"true"`
+	// ReadIdleTimeout defines the timeout after which a health check using ping frame will be carried out if no frame is received on the HTTP/2 connection.
+	// If zero, no health check is performed.
+	ReadIdleTimeout ptypes.Duration `json:"readIdleTimeout,omitempty" toml:"readIdleTimeout,omitempty" yaml:"readIdleTimeout,omitempty" export:"true"`
+	// PingTimeout defines the timeout after which the HTTP/2 connection will be closed if a response to ping is not received.
+	PingTimeout ptypes.Duration `json:"pingTimeout,omitempty" toml:"pingTimeout,omitempty" yaml:"pingTimeout,omitempty" export:"true"`
 }
 
-// SetDefaults sets the default values.
+// SetDefaults sets the default ForwardingTimeouts values.
 func (f *ForwardingTimeouts) SetDefaults() {
 	f.DialTimeout = ptypes.Duration(30 * time.Second)
-	f.IdleConnTimeout = ptypes.Duration(90 * time.Second)
+	f.IdleConnTimeout = DefaultIdleConnTimeout
 	f.PingTimeout = ptypes.Duration(15 * time.Second)
 }