### What does this PR do?

This pull request improves the HTTP performances between Traefik and the backends by using [FastHTTP](https://github.com/valyala/fasthttp) client primitives, for writing/reading HTTP requests as well as responses. The new HTTP proxy implementation using the FastHTTP proxy will only be used when making HTTP/1.1 requests to the backends.

This pull request also brings the following breaking changes:
- `passHostHeader` option is now a `ServersTransport` config option
- `responseForwarding.flushInterval` option has been removed as chunked responses are now handled automatically
- HTTP/2 is now disabled by default for communications between Traefik and the backends, unless the `enableHTTP2` option is set to `true` or the server URL uses the `h2c` scheme   
- To be consistent with other resources the `ServersTransport` configuration in the static configuration has been removed. It is now possible to configure a `default` ServersTransport as explained in the documentation.

### Benchmarks

Before:
```
Running 20s test @ http://traefik/bench
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.51ms    0.94ms  23.57ms   77.31%
    Req/Sec    17.44k     1.32k   51.46k    99.63%
  Latency Distribution
     50%    1.32ms
     75%    1.91ms
     90%    2.63ms
     99%    4.85ms
  1390186 requests in 20.10s, 135.23MB read
Requests/sec:  69166.15
Transfer/sec:      6.73MB
```

After:
```
Running 20s test @ http://traefik/bench
  4 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.12ms  675.32us  17.63ms   78.60%
    Req/Sec    23.28k     1.97k   77.69k    99.63%
  Latency Distribution
     50%    1.00ms
     75%    1.41ms
     90%    1.89ms
     99%    3.52ms
  1855455 requests in 20.10s, 180.49MB read
Requests/sec:  92314.15
Transfer/sec:      8.98MB
```

### Motivation

To have better performances.

### More

- [X] Added/updated tests
- [X] Added/updated documentation

### Additional Notes

Co-authored-by: Julien Salleyron <julien.salleyron@gmail.com>