@joeky888 2023-03-24
How about using [hertz](https://github.com/cloudwego/hertz) from ByteDance? It uses io_uring under the hood and supports http/2, performance is slightly better than fasthttp.