@mpl 2023-03-28
Hello @joeky888  ,

We haven't investigated/tried hertz because we're satisfied with what we have currently achieved with FastHTTP, and it would probably imply some additional significant effort to redo with hertz. So unless there's a really compelling argument to follow that lead, we're inclined to keep things as they are for now.
But thanks for the heads up nonetheless.