@joeky888 2023-03-28
Sure! We hope to see an ingress controller based on io_uring in the future. Hertz is still young, and [monoio](https://github.com/bytedance/monoio) is far from production-ready, both are developed by ByteDance.