@nmengin 2024-01-03
I move the feature to the `next` milestone.
According to the [v3 migration plan](https://github.com/traefik/traefik/issues/10260#issuecomment-1867474446), with the maintainers, we do not want to bring such a breaking feature in v3.0.

We prefer to postpone this feature, make it non-breaking, and embed it in a future v3 minor version.