@shkarface 2024-05-03
Now that v3 is officially released, @nmengin when do you think is a good time to look into this issue?