@nmengin 2024-05-27
Hello @shkarface,

Thank you for reaching out.
Indeed, we have planned to tackle this topic soon.

Note that this feature will be brought in a minor release, and it won't be enabled by default.