@kevinpollet 2024-09-24
Superseded by https://github.com/traefik/traefik/pull/11122